% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_chromosome_location.R
\name{calculate_chromosome_location}
\alias{calculate_chromosome_location}
\title{Calculate Chromosome Location}
\usage{
calculate_chromosome_location(
  data.site_proba_df,
  output_csv = NULL,
  output_plot = NULL
)
}
\arguments{
\item{data.site_proba_df}{m6anet output dataframe containing at least a column 'ensembl_gene_name'.}

\item{output_csv}{Optional path to save chromosome summary CSV.}

\item{output_plot}{Optional path to save barplot of chromosome distribution.}
}
\value{
chr_locations Merged data frame with chromosome info.
}
\description{
Adds chromosome annotation to a data frame of high-confidence m6A sites using \code{org.Hs.eg.db}.
}
\examples{
# Load example data
df_m6a <- read.csv(system.file("extdata", "sampleA_site_proba.csv", package = "m6ASeqTools"))

# Run chromosome location summary
df_with_chr <- calculate_chromosome_location(
  df_m6a,
  output_csv = tempfile(fileext = ".csv"),
  output_plot = tempfile(fileext = ".pdf")
)
head(df_with_chr)
}
