% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_weighted_mod_ratio.R
\name{calculate_weighted_mod_ratio}
\alias{calculate_weighted_mod_ratio}
\title{Calculate Weighted Modification Ratio}
\usage{
calculate_weighted_mod_ratio(data_frame, mod_ratio_column = "mod_ratio")
}
\arguments{
\item{data_frame}{A data frame with m6A site-level data with transcript regions (from \code{get_transcript_region_lengths()})}

\item{mod_ratio_column}{Character string, name of the column with modification ratios (default: "mod_ratio")}
}
\value{
A data frame with weighted modification ratios per transcript
}
\description{
Calculates the sum of a given modification ratio column for each transcript,
then normalizes it by transcript length.
}
