% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_m6a_distribution.R
\name{compare_m6a_distribution}
\alias{compare_m6a_distribution}
\title{Compare m6A Distribution between Samples}
\usage{
compare_m6a_distribution(named_list)
}
\arguments{
\item{named_list}{list of vectors (gene names, transcript ids, or site ids per group)}
}
\value{
a list of unique elements per group, pairwise intersections, common elements
to all, pairwise intersections excluding the third (if exactly 3 groups are input)
}
\description{
Compares common and unique m6A sites, transcripts, or genes for 2 or more groups
}
