% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_m6a_sites.R
\name{filter_m6a_sites}
\alias{filter_m6a_sites}
\title{Filter m6A Sites}
\usage{
filter_m6a_sites(data.site_proba_df, prob_modified = 0.9)
}
\arguments{
\item{data.site_proba_df}{The datadata frame from m6Anet output containing column 'probability_modified'}

\item{prob_modified}{The cut-off for probability modified. The default is 0.9 and above.}
}
\value{
filtered dataframe of m6A sites
}
\description{
This function filters m6A sites from m6anet output based on probability_modified criteria.
}
\examples{
filtered_sites <- filter_m6a_sites(data.site_proba_df, prob_modified=0.5)
}
