% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_transcript_region_lengths.R
\name{get_transcript_region_lengths}
\alias{get_transcript_region_lengths}
\title{Get Transcript Region Lengths}
\usage{
get_transcript_region_lengths(gtf_path)
}
\arguments{
\item{gtf_path}{The path to the GTF file of Gencode annotations.}
}
\value{
A data frame with transcript IDs and lengths/positions of UTRs and CDS regions.
}
\description{
This function calculates the lengths of the 5' UTR, CDS, and 3' UTR for each transcript
from a GTF annotation file using GenomicFeatures and summarizes them in a data frame.
}
\examples{
gtf_path <- 'gencode.v45.chr_patch_hapl_scaff.annotation.gtf'
tx_region <- get_transcript_region_lengths(gtf_path)
}
