% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log2fc_weighted_mod_ratio.R
\name{log2fc_weighted_mod_ratio}
\alias{log2fc_weighted_mod_ratio}
\title{Compare weighted mod ratio for two groups}
\usage{
log2fc_weighted_mod_ratio(
  group1_df,
  group2_df,
  group1_name = "Group1",
  group2_name = "Group2"
)
}
\arguments{
\item{group1_df}{Data frame for Group 1 (CONTROL), must include
\code{weighted_mod_ratio}}

\item{group2_df}{Data frame for Group 2 (EXPERIMENTAL), must include
\code{weighted_mod_ratio}}

\item{group1_name}{Name of Group 1 (used as suffix in output columns)}

\item{group2_name}{Name of Group 2 (used as suffix in output columns)}
}
\value{
A data frame with merged columns and \code{log2fc_weighted_mod_ratio}
}
\description{
Merges two data frames on commonly modified transcripts and calculates the
log2 fold change of the weighted modification ratio between groups.
}
\examples{
A_vs_B_log2fc_weighted_mod_ratio <- log2fc_weighted_mod_ratio(A_weighted_mod_df, B_weighted_mod_df, group1_name = "A", group2_name = "B")
}
