% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m6AnetAnalyzer_part1.R
\name{m6AnetAnalyzer_part1}
\alias{m6AnetAnalyzer_part1}
\title{Automate core steps of m6ASeqTools pipeline (Part 1)}
\usage{
m6AnetAnalyzer_part1(
  m6Aout,
  gtf_path,
  output_directory,
  mod_ratio_df,
  probability = 0.9
)
}
\arguments{
\item{m6Aout}{Data frame from m6Anet output (e.g., \code{data.site_proba.csv})}

\item{gtf_path}{File path to GTF annotation file (e.g., GENCODE)}

\item{output_directory}{Path where outputs will be saved}

\item{mod_ratio_df}{A string identifier used to name output files and the resulting object}

\item{probability}{Numeric threshold for high-confidence modification sites (default = 0.9)}
}
\value{
Writes multiple output files (CSV, PDF, HTML) to \code{output_directory} and assigns
the weighted modification ratio dataframe to the global environment
}
\description{
This function performs a series of common m6ASeqTools analyses including QC summary, filtering, region annotation,
chromosome location annotation, and weighted modification ratio calculation.
}
