% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_relative_tx_regions_to_m6a.R
\name{map_relative_tx_regions_to_m6A}
\alias{map_relative_tx_regions_to_m6A}
\title{Map Relative Transcript Regions to m6A Sites}
\usage{
map_relative_tx_regions_to_m6A(data.site_proba_df, df_transcript_lengths)
}
\arguments{
\item{data.site_proba_df}{The dataframe from m6Anet output containing columns
\code{ensembl_transcript_id} and \code{transcript_position}}

\item{df_transcript_lengths}{The output dataframe from
\code{get_transcript_region_lengths()} with transcript IDs and lengths/positions
of UTRs and CDS regions.}
}
\value{
A data frame with annotated transcript positions per m6A site,
relative to the region of the transcript
}
\description{
This function calculates the relative transcript regions of the 5' UTR, CDS,
and 3' UTR for each transcript using the region length df from
\code{get_transcript_region_lengths()}
}
