% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_m6anet_output.R
\name{summarize_m6anet_output}
\alias{summarize_m6anet_output}
\title{Summarize m6anet output}
\usage{
summarize_m6anet_output(
  data.site_proba_df,
  output_file = "m6anet_qc_report.html",
  verbose = TRUE,
  output_dir = NULL
)
}
\arguments{
\item{data.site_proba_df}{The 'datadata frame from m6Anet output.}

\item{output_file}{Name of the output HTML file.}

\item{verbose}{Whether you want the progress messages or not.}
}
\value{
Renders an HTML summary QC report.
}
\description{
This function summarizes the output of m6anet. This results in a html file with plots and
metrics. You can also put pooled data from m6anet as input.
}
