% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_weighted_mod_ratio_and_DGE.R
\name{summarize_weighted_mod_ratio_and_DGE}
\alias{summarize_weighted_mod_ratio_and_DGE}
\title{Summarize Weighted Modification Ratio and DGE and Provide annotations for clusters}
\usage{
summarize_weighted_mod_ratio_and_DGE(
  df,
  group1_name = "Group1",
  group2_name = "Group2"
)
}
\arguments{
\item{df}{Output from weighted_mod_ratio_and_DGE}

\item{group_1_name}{Name for group 1
@param group_2_name Name for group 2}
}
\value{
filtered dataframe of m6A sites
}
\description{
This function filters m6A sites from m6anet output based on probability_modified criteria.
}
