% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_wmr_dge_with_replicates.R
\name{summarize_wmr_dge_with_replicates}
\alias{summarize_wmr_dge_with_replicates}
\title{Summarize Weighted Modification Ratio and DGE and Provide annotations for clusters}
\usage{
summarize_wmr_dge_with_replicates(
  df,
  group1_name = "Group1",
  group2_name = "Group2",
  log2FC_dge,
  diff_wmr = "median_diff",
  dge_sig = "dge_sig",
  wmr_sig = "wmr_sig"
)
}
\arguments{
\item{df}{Output from weighted_mod_ratio_and_DGE}

\item{group1_name}{Name for group 1}

\item{group2_name}{Name for group 2}

\item{log2FC_dge}{Column name for log2 fold change (DGE)}

\item{diff_wmr}{Column name for the calculated difference in median weighted mod ratios}

\item{dge_sig}{Column name for gene log2fc significance boolean (e.g., adjusted p < 0.05)}

\item{wmr_sig}{Column name for differential test significance boolean (e.g., adjusted p < 0.05)}
}
\value{
A list containing filtered data frames and summaries
}
\description{
This function groups transcripts into functional categories based on gene expression and m6A methylation changes,
and summarizes transcript biotype and regional site distributions.
}
