% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_weighted_mod_ratio_and_DGE.R
\name{summarize_weighted_mod_ratio_and_DGE}
\alias{summarize_weighted_mod_ratio_and_DGE}
\title{Summarize Weighted Modification Ratio and DGE and Provide annotations for clusters}
\usage{
summarize_weighted_mod_ratio_and_DGE(
  df,
  group1_name = "Group1",
  group2_name = "Group2",
  log2FC_dge,
  log2fc_wmr,
  sig_col = "significant",
  combine = TRUE
)
}
\arguments{
\item{df}{Output from weighted_mod_ratio_and_DGE}

\item{group1_name}{Name for group 1}

\item{group2_name}{Name for group 2}

\item{log2FC_dge}{Column name for log2 fold change (DGE)}

\item{log2fc_wmr}{Column name for log2 fold change (weighted mod ratio)}

\item{sig_col}{Column name for significance boolean (e.g., adjusted p < 0.05)}

\item{combine}{Logical; whether to return combined summaries}
}
\value{
A list containing filtered data frames and summaries
}
\description{
This function groups transcripts into functional categories based on gene expression and m6A methylation changes,
and summarizes transcript biotype and regional site distributions.
}
